avoid error output from 'which', and substitute 'false' builtin
When none of the three yacc tools were installed, 'which' would print
the PATH to stderr three times. Now stderr is directed to /dev/null so
that 'which' may fail in silence.

Also, the trailing 'false' statement was redundant. If it were reached
the whole statement would be false anyway. Seems like the intent was to
have the result of the statement become a call to the builtin 'false'
instead of yacc when yacc isn't available. 'echo false' achieves that.